Struct ringbuffer::ConstGenericRingBuffer

pub struct ConstGenericRingBuffer<T, const CAP: usize> { /* private fields */ }

The ConstGenericRingBuffer struct is a RingBuffer implementation which does not require alloc but uses const generics instead.

ConstGenericRingBuffer allocates the ringbuffer on the stack, and the size must be known at compile time through const-generics.

Example

use ringbuffer::{ConstGenericRingBuffer, RingBuffer};

let mut buffer = ConstGenericRingBuffer::<_, 2>::new();

// First entry of the buffer is now 5.
buffer.push(5);

// The last item we pushed is 5
assert_eq!(buffer.back(), Some(&5));

// Second entry is now 42.
buffer.push(42);

assert_eq!(buffer.peek(), Some(&5));
assert!(buffer.is_full());

// Because capacity is reached the next push will be the first item of the buffer.
buffer.push(1);
assert_eq!(buffer.to_vec(), vec![42, 1]);

Implementations

impl<T, const CAP: usize> ConstGenericRingBuffer<T, CAP>

pub const fn new<const N: usize>() -> Selfwhere ConstGenericRingBuffer<T, CAP>: From<ConstGenericRingBuffer<T, N>>,

Creates a const generic ringbuffer, size is passed as a const generic.

Note that the size does not have to be a power of two, but that not using a power of two might be significantly (up to 3 times) slower.